v0.2.0

+++ src/main.rs
@@ -587,7 +587,7 @@ mod tests {
         let read_qualities: Vec<u64> = vec![10, 20, 20, 30];
         let read_lengths: Vec<u64> = vec![10, 20, 20, 30];
         let keep_percent: f64 = 0.25;
-        let keep_bases: u64 = 0;
+        let keep_bases: usize = 0;
         let indices = retain_indexed_quality_reads(read_qualities, read_lengths, keep_percent, keep_bases).unwrap();
 
         assert_eq!(indices, None)
@@ -599,7 +599,7 @@ mod tests {
         let read_qualities: Vec<u64> = vec![10, 20, 20, 30];
         let read_lengths: Vec<u64> = vec![10, 20, 20, 30];
         let keep_percent: f64 = 1.0;
-        let keep_bases: u64 = 30;
+        let keep_bases: usize = 30;
         let indices = retain_indexed_quality_reads(read_qualities, read_lengths, keep_percent, keep_bases).unwrap();
 
         assert_eq!(indices, None)
@@ -611,7 +611,7 @@ mod tests {
         let read_qualities: Vec<u64> = vec![10, 20, 20, 30];
         let read_lengths: Vec<u64> = vec![10, 20, 20, 30];
         let keep_percent: f64 = 0.25;
-        let keep_bases: u64 = 30;
+        let keep_bases: usize = 30;
         let indices = retain_indexed_quality_reads(read_qualities, read_lengths, keep_percent, keep_bases).unwrap();
 
         assert_eq!(indices, None)
